Moved slack widget and aligned boxes
