+++ frontend/src/app/navigation/navigation.component.ts
@@ -11,6 +11,7 @@ import { ActivatedRoute, Router, RouterEvent } from '@angular/router';
 import { Profile, ProfileService } from '../profile/profile.service';
 import { PermissionService } from '../permission.service';
 import { NagivationAdminGearService } from './navigation-admin-gear.service';
+import { SlackInviteBox } from '../shared/slack-invite-box/slack-invite-box.widget';
 
 @Component({
   selector: 'app-navigation',
@@ -36,7 +37,8 @@ export class NavigationComponent implements OnInit, OnDestroy {
     private breakpointObserver: BreakpointObserver,
     protected navigationService: NavigationTitleService,
     protected navigationAdminGearService: NagivationAdminGearService,
-    protected errorDialog: MatDialog
+    protected errorDialog: MatDialog,
+    protected slackDialog: MatDialog
   ) {
     this.profile$ = this.profileService.profile$;
     this.checkinPermission$ = this.permission.check(
@@ -71,6 +73,14 @@ export class NavigationComponent implements OnInit, OnDestroy {
     if (this.isHandset) {
       nav.close();
     }
+  }
+
+  onSlackInviteClick(): void {
+    const dialogRef = this.slackDialog.open(SlackInviteBox, {
+      width: '1000px',
+      autoFocus: false
+    });
+    dialogRef.afterClosed().subscribe();
   }
 
   private initErrorDialog() {
