+++ frontend/src/app/shared/slack-invite-box/slack-invite-box.widget.ts
